fix add contact

+++ Zalo_Client/src/services/contactService.js
@@ -2,7 +2,7 @@ const axios = require('axios');
 const http = require('../controllers/http');
 
 class ContactService {
-     addNew(currentUserId, contactId) {
+    addNew(currentUserId, contactId) {
         return new Promise(async (resolve, reject) => {
             let contactExists = await this.checkExistsContact(currentUserId, contactId);
             if (contactExists) {
@@ -12,8 +12,8 @@ class ContactService {
                     userId: currentUserId,
                     contactId: contactId
                 };
-            const newContact = await axios.post(http + '/contacts', newContactItem);
-               return resolve(newContact);
+                const newContact = await axios.post(http + '/contacts', newContactItem);
+                return resolve(newContact);
             }
         })
     }
@@ -21,18 +21,23 @@ class ContactService {
     remove(currentUserId, contactId) {
         return new Promise(async (resolve, reject) => {
             let findContact = await this.checkExistsContact(currentUserId, contactId);
-            let removeContact = await axios.delete(http + '/contacts/' + findContact._id);
-            if(removeContact.data == true){
-                return resolve(true);
-            }else{
-                return reject(false);
+            if(findContact){
+                await axios.delete(http + '/contacts/' + findContact._id)
+                .then(resolve(true))
+                .catch(reject(false));
             }
+           
+            // if (removeContact.data == true) {
+            //     return resolve(true);
+            // } else {
+            //     return reject(false);
+            // }
         })
     }
 
     async checkExistsContact(userId, contactId) {
-       let checkExists = await axios.get(http + '/contacts/search/' + userId + '/' + contactId)
-      return checkExists.data.contact;
+        let checkExists = await axios.get(http + '/contacts/search/' + userId + '/' + contactId)
+        return checkExists.data.contact;
     }
 }
